requirements.txt
fastapi
uvicorn
python-multipart
torch
opencv-python-headless
numpy
Pillow
transformers
